Refactor to isolate the two tests from each other
